+++ lib/shared/link_preview_card.dart
@@ -256,9 +256,6 @@ class LinkPreviewCard extends StatelessWidget {
   }
 
   void triggerOnTap(BuildContext context) async {
-    final ThunderState state = context.read<ThunderBloc>().state;
-    final openInExternalBrowser = state.openInExternalBrowser;
-
     if (isUserLoggedIn && markPostReadOnMediaView) {
       // Mark post as read when on the feed page
       try {
@@ -296,7 +293,9 @@ class LinkPreviewCard extends StatelessWidget {
         }
       }
 
-      openLink(context, url: originURL!, openInExternalBrowser: openInExternalBrowser);
+      if (context.mounted) {
+        handleLink(context, url: originURL!);
+      }
     }
   }